Delete Size2i

+++ src/main/kotlin/at/orchaldir/gm/app/html/model/Size2d.kt
@@ -4,14 +4,14 @@ import at.orchaldir.gm.app.HEIGHT
 import at.orchaldir.gm.app.WIDTH
 import at.orchaldir.gm.app.html.field
 import at.orchaldir.gm.app.parse.combine
-import at.orchaldir.gm.utils.math.Size2i
+import at.orchaldir.gm.utils.math.Size2d
 import at.orchaldir.gm.utils.math.unit.SiPrefix
 import io.ktor.http.*
 import kotlinx.html.HtmlBlockTag
 
 // show
 
-fun HtmlBlockTag.fieldSize(name: String, size: Size2i) {
+fun HtmlBlockTag.fieldSize(name: String, size: Size2d) {
     field(name, "${size.width} x ${size.height}")
 }
 
@@ -19,7 +19,7 @@ fun HtmlBlockTag.fieldSize(name: String, size: Size2i) {
 
 fun HtmlBlockTag.selectSize(
     param: String,
-    size: Size2i,
+    size: Size2d,
     minValue: Long,
     maxVale: Long,
     prefix: SiPrefix,
@@ -35,7 +35,7 @@ fun parseSize(
     parameters: Parameters,
     param: String,
     prefix: SiPrefix,
-) = Size2i(
+) = Size2d(
     parseDistance(parameters, combine(param, WIDTH), prefix, 1),
     parseDistance(parameters, combine(param, HEIGHT), prefix, 1),
 )
